+++ src/index.js
@@ -8,6 +8,7 @@
  */
 'use strict';
 
+export * from './timeline';
 export {default as Timeline} from './timeline';
 export * from './types';
 
@@ -18,7 +19,7 @@ export {default as ItemRenderer} from './components/ItemRenderer';
 export {Marker} from './components/Marker';
 export {default as Selectbox} from './components/selector';
 export {BackgroundLayer} from './components/BackgroundLayer';
-export {HighlightedInterval} from './components/HighlightedInterval';
+export * from './components/HighlightedInterval';
 
 // consts
 export {timebarFormat} from './consts/timebarConsts';