+++ CHANGELOG.md
@@ -1,21 +1,38 @@
-## Unreleased (probably 2.4.0)
+## Unreleased
+
+* Internal: for `demo-app`, upgraded storybook version to v7.5. It also uses vite now.
+* Small issue: on props changed, the width of the table was reset to the original value.
+* [Highlighted interval positioned at the beginning of the gantt was not displayed](https://github.com/flower-platform/react-timeline-10000/pull/70)
+
+## v2.5.0
+
+* [Separated lib and demo-app](https://github.com/flower-platform/react-timeline-10000/commit/b5936969f7ffe67a52139c04b2275849e9c5b078)
+* [Upgraded to TAD v0.11.0](https://github.com/flower-platform/react-timeline-10000/commit/a87a58b74898e3a5017b8e4d5028e9e5a4c0f325), generated screenshots, .md reports, and first .md docs 
+
+## v2.4.1
+
+### Fixed
+
+* [Reseting start/end date didn't work](https://github.com/flower-platform/react-timeline-10000/pull/65)
+
+## v2.4.0
 
 ### Added
 
 * [Horizontal scrollbar](https://github.com/flower-platform/react-timeline-10000/pull/57)
-* [Zoom](https://github.com/flower-platform/react-timeline-10000/pull/54)
-
-# v2.3.0-SNAPSHOT2
-
-### Added
 * [Added `onTableResize` property](https://github.com/flower-platform/react-timeline-10000/pull/62)
-
-# v2.3.0-SNAPSHOT1
 
 ### Fixed
 
 * [External drag and drop doesn't work](https://github.com/flower-platform/react-timeline-10000/pull/60)
 * [Changing the `width` property of the table didn't updated of the UI](https://github.com/flower-platform/react-timeline-10000/pull/59)
+* [Various small isues](https://github.com/flower-platform/react-timeline-10000/pull/63)
+  * Problem with the scroll after reseting the `groups`
+  * `SplitPanel` has `absolute` position instead of `relative`position 
+  * Exception when dragging above the timeline
+  * Exception when reseting the `groups` of a gantt that was scrolled
+  * Gantt surface was not receiving events because of the `backgroundLayers`
+  * Background layers were overlapping the vertical scrollbar
 
 ## v2.3.0